update the name of the stderr output and clean up
This prevents anyone who has a .gitignore file catching any *.log files from having issues where the PR isn't created when the release-plan fails. Essentially the peter-evans/create-pull-request action needs a file change to open the PR and we leave the release-plan-stderr.txt in the output when we have a failure to make sure the PR is created.

This PR also make sure to clean up the release-plan-stderr.txt when the plan succeeds